+++ packages/site/.eslintrc.cjs
@@ -1,0 +1,12 @@
+/** @type {import("eslint").Linter.Config} */
+module.exports = {
+  extends: ['@cofhe/eslint-config/library'],
+  parser: '@typescript-eslint/parser',
+  parserOptions: {
+    project: true,
+  },
+  ignorePatterns: [
+    'scripts/hardhat-typecheck/.generated/',
+    '**/*.mdx',
+  ],
+};